Build: Remove causes of pedantic warnings. Addresses issue #352 and PR #242.

+++ glslang/MachineIndependent/Scan.h
@@ -173,7 +173,7 @@ public:
         loc[getLastValidSourceIndex()].name = filename;
     }
 
-    void setFile(const char* filename, size_t i)
+    void setFile(const char* filename, int i)
     {
         if (i == getLastValidSourceIndex()) {
             logicalSourceLoc.name = filename;
